+++ src/components/node-list/filter-nodes.test.js
@@ -1,0 +1,25 @@
+import { getNodeIDs } from './filter-nodes';
+
+describe('filter-nodes', () => {
+  describe('getNodeIDs', () => {
+    const generateNodes = (type, count) =>
+      Array.from(new Array(count)).map((d, i) => ({
+        id: type + i
+      }));
+
+    const nodes = {
+      data: generateNodes('data', 10),
+      task: generateNodes('task', 10),
+      parameters: generateNodes('parameters', 10)
+    };
+
+    it('returns a list of node IDs', () => {
+      const nodeIDs = getNodeIDs(nodes);
+      expect(nodeIDs).toHaveLength(30);
+      expect(nodeIDs).toEqual(expect.arrayContaining([expect.any(String)]));
+      expect(nodeIDs).toContain('data0');
+      expect(nodeIDs).toContain('task1');
+      expect(nodeIDs).toContain('parameters1');
+    });
+  });
+});
